feat(refactor): Add thunk to make api req for zendesk
Added the submitConfirm content to a easy peasy thunk.
